[qcommon] Make bit computation well-defined by computing everything as unsigned ints.

+++ engine/code/qcommon/qcommon.h
@@ -195,7 +195,7 @@ void		NET_Sleep(int msec);
 						// will overflow the reliable commands buffer
 #define MAX_DOWNLOAD_BLKSIZE		1024	// 896 byte block chunks
 
-#define NETCHAN_GENCHECKSUM(challenge, sequence) ((challenge) ^ ((sequence) * (challenge)))
+#define NETCHAN_GENCHECKSUM(challenge, sequence) ((int)((unsigned int)(challenge) ^ ((unsigned int)(sequence) * (unsigned int)(challenge))))
 
 /*
 Netchan handles packet fragmentation and out of order / duplicate suppression
